Remove conflict on email.py , Added html content emails service
