+++ .env_sample
@@ -15,3 +15,4 @@ SMTP_USERNAME=
 SMTP_PASSWORD=
 REPLY_TO_EMAIL=
 FROM_EMAIL=
+TARGET_EMAIL_ADDRESS='test@gmail.com'
