+++ requirements.txt
@@ -2,4 +2,5 @@ pandas==2.2.0
 pyarrow==15.0.0
 polars==0.20.10
 s3fs==2024.2.0 
+s3pathlib==2.1.2
 boto-session-manager==1.7.2
